final comments
