+++ perf/_precision.py
@@ -40,6 +40,24 @@ def compare(a: Any, b: Any, relative: bool = False) -> float:
             raise
         raise TypeError(f"Inputs must both be tensors, dicts, or iterables of tensors. Got types: {type(a)} and {type(b)}")
 
+def get_violation_pct(gold, ref, test, tol, atol=0):
+    """Calculate the violation pct of the difference between (ref-gold) and (test-gold)
+
+    Args:
+        gold: Gold tensor or iterable of tensors
+        ref: Reference tensor or iterable of tensors
+        test: Test tensor or iterable of tensors
+        tol: float. Tolerance for difference
+        atol: float. Absolute tolerance for difference
+    """
+    gold_ref_diff = torch.abs(gold - ref)
+    gold_test_diff = torch.abs(gold - test)
+    abs_error = (gold_test_diff - ((1 + tol) * gold_ref_diff + atol))
+    violations = (abs_error > 0)
+    num_violations = violations.sum().item()
+    violation_pct = num_violations / gold.numel()
+    return violation_pct
+
 def measure_forward_precision(ref_fn: callable, fn: callable, ref_inputs: Dict[str, torch.Tensor], test_inputs: Dict[str, torch.Tensor], relative: bool) -> float:
     """Measure precision difference between reference and test implementations on forward pass.
